requirements.txt
django==1.11
gunicorn
pillow
psycopg2==2.7.3.2
django-jet
djangorestframework
django-rest-swagger
static3==0.5.1
django-storages
boto==2.42.0
django_cors_headers==2.1.0
boto3==1.5.27
